requirements.txt
python-telegram-bot==13.0
urllib3==1.26.15
requests==2.32.4
environs==14.2.0
standard_imghdr==3.13.0
